+++ app/interactors/send_reset_token.rb
@@ -1,0 +1,8 @@
+# Send reset token to user's email
+class SendResetToken
+  include Interactor
+
+  def call
+    ResetPasswordEmailJob.perform_later(user_id: context.user.id)
+  end
+end