+++ app/models/character_class.rb
@@ -5,4 +5,15 @@ class CharacterClass < ApplicationRecord
   has_many :combinateables, through: :combinations, source_type: 'Role'
   has_many :role_combinations, -> { where combinateable_type: 'Role' }, class_name: 'Combination'
   has_many :available_roles, through: :role_combinations, source: :combinateable, source_type: 'Role'
+
+  def to_hash
+    {
+      id.to_s => {
+        'name' => name,
+        'roles' => combinateables.inject({}) do |roles, role|
+          roles.merge(role.to_hash)
+        end
+      }
+    }
+  end
 end