+++ db/migrate/20191124214606_add_reset_token_to_users.rb
@@ -1,0 +1,6 @@
+class AddResetTokenToUsers < ActiveRecord::Migration[5.2]
+  def change
+    add_column :users, :reset_password_token, :string
+    add_column :users, :reset_password_token_sent_at, :datetime
+  end
+end