+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2019_11_24_121827) do
+ActiveRecord::Schema.define(version: 2019_11_24_214606) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pg_trgm"
@@ -304,6 +304,8 @@ ActiveRecord::Schema.define(version: 2019_11_24_121827) do
     t.string "remember_digest"
     t.string "confirmation_token"
     t.datetime "confirmed_at"
+    t.string "reset_password_token"
+    t.datetime "reset_password_token_sent_at"
     t.index ["email"], name: "index_users_on_email", unique: true
   end