+++ frontend/components/flashes/flashes.js
@@ -1,0 +1,7 @@
+import "./flashes.scss";
+
+const $ = require("jquery")
+
+$('.content').on('click', '.flash-messages .alert button', function() {
+  $(this).parent().hide()
+})