+++ spec/policies/guild_policy_spec.rb
@@ -11,6 +11,44 @@ describe GuildPolicy do
   let!(:character3) { create :character, guild: guild3, user: user }
   let!(:guild_role3) { create :guild_role, guild: guild3 }
 
+  describe '#edit?' do
+    context 'for admin' do
+      let(:policy) { described_class.new(guild3, user: admin) }
+
+      it 'returns false' do
+        expect(policy_access).to eq false
+      end
+    end
+
+    context 'if user has gm character' do
+      let(:policy) { described_class.new(guild1, user: user) }
+
+      it 'returns true' do
+        expect(policy_access).to eq true
+      end
+    end
+
+    context 'if user has character with no gm or rl role' do
+      let(:policy) { described_class.new(guild2, user: user) }
+
+      it 'returns false' do
+        expect(policy_access).to eq false
+      end
+    end
+
+    context 'if user has character without role' do
+      let(:policy) { described_class.new(guild3, user: user) }
+
+      it 'returns false' do
+        expect(policy_access).to eq false
+      end
+    end
+
+    def policy_access
+      policy.edit?
+    end
+  end
+
   describe '#management?' do
     context 'for admin' do
       let(:policy) { described_class.new(guild3, user: admin) }