+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- reset password process
+- amount of approved characters for event
+
 ### Modified
 - responsive navigation
+- tests
 
 ### Fixed
 - fix static management page